+++ packages/openapi-generator/src/options/flags.ts
@@ -94,24 +94,8 @@ export const generatorFlags = {
     char: 'c',
     parse: input => resolve(input),
     description:
-      'Set the path to the a file containing the options for generation. If the config flag is used, no other flags can be used. If a directory is passed, a `config.json` file is read from this directory.',
+      'Set the path to the a file containing the options for generation. If other flags are used, they overwrite the options set in the config. If a directory is passed, a `config.json` file is read from this directory.',
     helpValue: '<path/to/config.json>',
-    required: false,
-    exclusive: [
-      'input',
-      'outputDir',
-      'transpile',
-      'include',
-      'overwrite',
-      'clearOutputDir',
-      'skipValidation',
-      'tsConfig',
-      'packageJson',
-      'verbose',
-      'optionsPerService',
-      'packageVersion',
-      'readme',
-      'metadata'
-    ]
+    required: false
   })
 };